+++ src/components/ConfigUi/Reporting/Widget/index.js
@@ -1,3 +1,7 @@
+/*
+ * Copyright © 2015-2020 Serenova, LLC. All rights reserved.
+ */
+
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
@@ -5,40 +9,49 @@ import styled from 'styled-components';
 const DEFAULT_COLORS = { info: '#aaa', success: '#54b84f', warning: '#f58c00', error: 'red' };
 
 const Label = styled.h3`
-  font-size: 1.5em;
+  font-size: 21px;
+  font-weight: 700;
+  color: #474747;
 `;
 
 const Value = styled.p`
-  font-size: 3em;
+  font-size: 21px;
+  font-weight: 700;
   margin: 0.25em auto;
   color: ${({ color, componentType }) => color || DEFAULT_COLORS[componentType]};
 `;
 
 const Wrapper = styled.div`
-  padding: 0.5em;
+  padding: 1em;
   text-align: center;
   display: flex;
   flex-direction: column;
   justify-content: center;
   height: 100%;
+  border: 1px solid lightgray;
 `;
 
-function Stat({ label, value, componentType = 'info', color }) {
+function Widget({ className, label, value, componentType = 'info', color, children }) {
   return (
-    <Wrapper className={`stat stat--${componentType}`}>
-      <Label className="stat__label">{label}</Label>
-      <Value className="stat__value" color={color} componentType={componentType}>
-        {value}
-      </Value>
+    <Wrapper className={className}>
+      <Label>{label}</Label>
+      {children && children}
+      {value && (
+        <Value color={color} componentType={componentType}>
+          {value}
+        </Value>
+      )}
     </Wrapper>
   );
 }
 
-Stat.propTypes = {
+Widget.propTypes = {
   label: PropTypes.string.isRequired,
-  value: PropTypes.string.isRequired,
+  value: PropTypes.string,
   componentType: PropTypes.oneOf(['info', 'success', 'warning', 'error']),
   color: PropTypes.string,
+  className: PropTypes.string,
+  children: PropTypes.element,
 };
 
-export default Stat;
+export default Widget;